refactor: seperate stub-components and patch-render
